Restore original name for binary compatibility
